arc-gltr-backend/requirements.txt
# pytorch-pretrained-bert>=0.6.1
transformers
torch
numpy
pandas
connexion[swagger-ui]
flask
PyYaml
flask_cors
docx2txt
PyPDF2
gunicorn
python-docx
